edit utilities
