+++ src/pages/PostPage.tsx
@@ -17,11 +17,7 @@ import { Linkify } from "../utils/utilities";
 import { fetchSinglePost } from "../redux/post/post.actions";
 import { useParams } from "react-router";
 import { SubmitHandler, useForm } from "react-hook-form";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { newComment } from "../redux/post/post.actions";
-import { theme } from "../styles/theme";
-
-const defaultTheme = createTheme(theme);
 
 export const PostPage = (): JSX.Element => {
   const dispatch = useAppDispatch();
@@ -50,102 +46,100 @@ export const PostPage = (): JSX.Element => {
   }, [dispatch, id]);
 
   return (
-    <ThemeProvider theme={defaultTheme}>
-      <Container component="main" maxWidth="md" sx={{ mt: 12 }}>
-        {posts.loading ? (
-          <CircularProgress />
-        ) : posts.singlePost.title ? (
-          <Box
-            sx={{
-              marginTop: 8,
-              display: "flex",
-              flexDirection: "column",
-              alignItems: "center",
-            }}
-            component="form"
-            onSubmit={handleSubmit(onSubmit)}
-            noValidate
-          >
-            <Grid container spacing={2}>
-              <Grid item xs={12}>
-                <Card>
-                  <CardContent>
-                    <Typography sx={{ fontWeight: "bold", fontSize: 18 }}>
-                      {posts.singlePost.title}
-                    </Typography>
-                    <Linkify sx={{ my: 1 }}>{posts.singlePost.text}</Linkify>
-                    <Typography sx={{ fontSize: 14 }}>
-                      posted at {posts.singlePost.createdAt.slice(11, 16)} on{" "}
-                      {posts.singlePost.createdAt.slice(0, 10)}
-                    </Typography>
-                    <Typography sx={{ fontSize: 14 }}>
-                      by {posts.singlePost.user.username}
-                    </Typography>
-                  </CardContent>
-                </Card>
-              </Grid>
-              {posts.singlePost.comments.length > 0 ? (
-                <Grid item xs={12}>
-                  <Typography sx={{ ml: 1, mb: 2 }}>Comments</Typography>
-                  {posts.singlePost.comments.toReversed().map((comment) => (
-                    <Card key={comment._id} sx={{ my: 1 }}>
-                      <CardContent key={comment.id}>
-                        <Linkify sx={{ mb: 1 }}>{comment.text}</Linkify>
-                        <Typography sx={{ fontSize: 14 }}>
-                          posted at {comment.createdAt.slice(11, 16)} on{" "}
-                          {comment.createdAt.slice(0, 10)}
-                        </Typography>
-                        <Typography sx={{ fontSize: 14 }}>
-                          by {comment.user.username}
-                        </Typography>
-                      </CardContent>
-                    </Card>
-                  ))}
-                </Grid>
-              ) : (
-                <Typography sx={{ m: 2, ml: 3 }}>
-                  Be the first to leave a comment!
-                </Typography>
-              )}
-              {auth.loggedInUser.access_token ? (
-                <Grid item xs={12}>
-                  <TextField
-                    {...register("comment", {
-                      required: "Please add a comment and try again",
-                      maxLength: {
-                        value: 240,
-                        message:
-                          "Comments cannot exceed 240 characters in length",
-                      },
-                    })}
-                    id="comment"
-                    label="Comment"
-                    variant="outlined"
-                    fullWidth
-                  />
-                  {errors.comment && (
-                    <Typography variant="caption" color="error">
-                      {errors.comment.message}
-                    </Typography>
-                  )}
-                  <Button
-                    type="submit"
-                    variant="contained"
-                    sx={{ width: 90, mt: 2, mb: 10 }}
-                  >
-                    Submit
-                  </Button>
-                </Grid>
-              ) : (
-                <></>
-              )}
+    <Container component="main" maxWidth="md" sx={{ mt: 12 }}>
+      {posts.loading ? (
+        <CircularProgress />
+      ) : posts.singlePost.title ? (
+        <Box
+          sx={{
+            marginTop: 8,
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "center",
+          }}
+          component="form"
+          onSubmit={handleSubmit(onSubmit)}
+          noValidate
+        >
+          <Grid container spacing={2}>
+            <Grid item xs={12}>
+              <Card>
+                <CardContent>
+                  <Typography sx={{ fontWeight: "bold", fontSize: 18 }}>
+                    {posts.singlePost.title}
+                  </Typography>
+                  <Linkify sx={{ my: 1 }}>{posts.singlePost.text}</Linkify>
+                  <Typography sx={{ fontSize: 14 }}>
+                    posted at {posts.singlePost.createdAt.slice(11, 16)} on{" "}
+                    {posts.singlePost.createdAt.slice(0, 10)}
+                  </Typography>
+                  <Typography sx={{ fontSize: 14 }}>
+                    by {posts.singlePost.user.username}
+                  </Typography>
+                </CardContent>
+              </Card>
             </Grid>
-          </Box>
-        ) : (
-          <Typography>Post not found</Typography>
-        )}
-      </Container>
-    </ThemeProvider>
+            {posts.singlePost.comments.length > 0 ? (
+              <Grid item xs={12}>
+                <Typography sx={{ ml: 1, mb: 2 }}>Comments</Typography>
+                {posts.singlePost.comments.toReversed().map((comment) => (
+                  <Card key={comment._id} sx={{ my: 1 }}>
+                    <CardContent key={comment.id}>
+                      <Linkify sx={{ mb: 1 }}>{comment.text}</Linkify>
+                      <Typography sx={{ fontSize: 14 }}>
+                        posted at {comment.createdAt.slice(11, 16)} on{" "}
+                        {comment.createdAt.slice(0, 10)}
+                      </Typography>
+                      <Typography sx={{ fontSize: 14 }}>
+                        by {comment.user.username}
+                      </Typography>
+                    </CardContent>
+                  </Card>
+                ))}
+              </Grid>
+            ) : (
+              <Typography sx={{ m: 2, ml: 3 }}>
+                Be the first to leave a comment!
+              </Typography>
+            )}
+            {auth.loggedInUser.access_token ? (
+              <Grid item xs={12}>
+                <TextField
+                  {...register("comment", {
+                    required: "Please add a comment and try again",
+                    maxLength: {
+                      value: 240,
+                      message:
+                        "Comments cannot exceed 240 characters in length",
+                    },
+                  })}
+                  id="comment"
+                  label="Comment"
+                  variant="outlined"
+                  fullWidth
+                />
+                {errors.comment && (
+                  <Typography variant="caption" color="error">
+                    {errors.comment.message}
+                  </Typography>
+                )}
+                <Button
+                  type="submit"
+                  variant="contained"
+                  sx={{ width: 90, mt: 2, mb: 10 }}
+                >
+                  Submit
+                </Button>
+              </Grid>
+            ) : (
+              <></>
+            )}
+          </Grid>
+        </Box>
+      ) : (
+        <Typography>Post not found</Typography>
+      )}
+    </Container>
   );
 };
