--- src/utils/utilities.tsx
@@ -1,16 +1,0 @@
-export const Linkify = ({ children }) => {
-  const isUrl = (word) => {
-    const urlPattern =
-      /^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([-.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$/gm;
-    return word.match(urlPattern);
-  };
-
-  const addMarkup = (word) => {
-    return isUrl(word) ? `<a href="${word}">${word}</a>` : word;
-  };
-
-  const words = children.split(" ");
-  const formatedWords = words.map((w) => addMarkup(w));
-  const html = formatedWords.join(" ");
-  return <span dangerouslySetInnerHTML={{ __html: html }} />;
-};
